add image and file struct
