+++ lib/fnc/app.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:les3/UI/ProfilePage.dart';
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: ProfilePage(),
+    );
+  }
+}
