+++ test/widget_test.dart
@@ -7,6 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:les3/fnc/app.dart';
 
 import 'package:les3/main.dart';
